expose mergeWithComponents
