+++ packages/vulcan-forms/lib/components/FormComponent.jsx
@@ -1,12 +1,11 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Components } from 'meteor/vulcan:core';
-import { registerComponent } from 'meteor/vulcan:core';
+import { registerComponent, mergeWithComponents } from 'meteor/vulcan:core';
 import get from 'lodash/get';
 import isEqual from 'lodash/isEqual';
 import SimpleSchema from 'simpl-schema';
 import { isEmptyValue, getNullValue } from '../modules/utils.js';
-import mergeWithComponents from '../modules/mergeWithComponents';
 
 class FormComponent extends Component {
   constructor(props) {
