+++ packages/vulcan-forms/lib/components/FormGroup.jsx
@@ -2,8 +2,7 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { Components, Utils } from 'meteor/vulcan:core';
 import classNames from 'classnames';
-import { registerComponent } from 'meteor/vulcan:core';
-import mergeWithComponents from '../modules/mergeWithComponents';
+import { registerComponent, mergeWithComponents } from 'meteor/vulcan:core';
 
 const FormGroupHeader = ({ toggle, collapsed, label }) => (
   <div className="form-section-heading" onClick={toggle}>
