+++ packages/vulcan-forms/test/index.js
@@ -1,3 +1,2 @@
 import './schema_utils.test.js';
 import './components.test.js';
-import './mergeWithComponents.test.js';
