+++ packages/vulcan-lib/test/index.js
@@ -1,1 +1,2 @@
 import './handleOptions.test.js';
+import './components.test.js';
